backend/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
distlib==0.3.4
Django==4.0
djangorestframework==3.12.4
filelock==3.4.0
pipenv==2021.11.23
platformdirs==2.4.0
six==1.16.0
sqlparse==0.4.2
virtualenv==20.10.0
virtualenv-clone==0.5.7
